# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion Medical",
        "description": "Documentation API pour la  gestion Medical",
        "contact": {
            "email": "diopawa425@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Se connecter",
                "description": "Se connecter.",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Auth"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Auth"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/password/change": {
            "post": {
                "tags": [
                    "reset Password"
                ],
                "summary": "Reset Password",
                "description": "Reinitialiser votre mot de pass.",
                "operationId": "resetPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPassword"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResetPassword"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/consultations": {
            "get": {
                "tags": [
                    "dossier_medical & Consultation"
                ],
                "summary": "liste consultation",
                "description": "Liste de tous les consultations.",
                "operationId": "listConsultation",
                "responses": {
                    "200": {
                        "description": "Données récupérées avec succès.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/patients/create-consultation": {
            "post": {
                "tags": [
                    "dossier_medical & Consultation"
                ],
                "summary": "Consulter, mettre a jour dossier patient",
                "description": "Consulter et modifier dossier medical d'un patient.",
                "operationId": "consulterPatient",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Consultation"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Consultation"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/patients/consultation/{id}/update": {
            "put": {
                "tags": [
                    "dossier_medical & Consultation"
                ],
                "summary": "Modifier les infos d'une consultation",
                "description": "Modifier les infos d'une consultation.",
                "operationId": "updateConsultation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID consultation à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Consultation"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateur mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Personnel mis à jour avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Consultation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Auth": {
                "title": "Login",
                "description": "Schéma pour s'authentifier",
                "properties": {
                    "email": {
                        "description": "Login",
                        "type": "string",
                        "example": "user1@hopital.com"
                    },
                    "password": {
                        "description": "Mot de passe",
                        "type": "string",
                        "example": "passer"
                    }
                },
                "type": "object"
            },
...
            "Consultation": {
                "title": "Consultation",
                "description": "Schéma représentant un Consultation",
                "properties": {
                    "libelle": {
                        "description": "Nom de la consultation",
                        "type": "string",
                        "example": "Ophtalmologie"
                    },
                    "notes": {
                        "description": "Note prise lors de la consultation",
                        "type": "string",
                        "example": "mettez les notes ici"
                    },
                    "dossierMedical": {
                        "$ref": "#/components/schemas/DossierMedical"
                    },
                    "medecin_id": {
                        "description": "Medecin associée",
                        "type": "integer",
                        "example": "59"
                    },
                    "patient_id": {
                        "description": "Patient consulté",
                        "type": "integer",
                        "example": "1"
                    },
                    "service_id": {
                        "description": "Service a aller",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "DossierMedical": {
                "description": "Dossier médical du patient",
                "properties": {
                    "antecedents": {
                        "description": "Liste des antécédents du patient",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Diabète"
                        }
                    },
                    "diagnostics": {
                        "description": "Liste des diagnostics du patient",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Appendicite"
                        }
                    },
                    "traitements": {
                        "description": "Liste des traitements du patient",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Antibiotiques"
                        }
                    },
                    "prescriptions": {
                        "description": "Liste des prescriptions médicales",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "Paracétamol 500mg"
                        }
                    }
                },
                "type": "object"
            },
...
            "ResetPassword": {
                "title": "Reset Password",
                "description": "Changer votre mot de passe ",
                "properties": {
                    "password": {
                        "description": "Reset Password",
                        "type": "string",
                        "example": "password123"
                    },
                    "password_confirmation": {
                        "description": "Confirm Password",
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Entrez le token Sanctum reçu après connexion.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}